#1: Feat hero section (#1)
#1: Feat hero section (#1)

What changed?
====================
This commit adds the hero section to the index of the website. It uses next-themes to toggle between light and dark mode, with a sun and moon at the top acting as the toggle button.

Why?
====================
To allow users to view a short description of me and add animations for fun user interactivity.